Added duplicate packet recognisation
Added a way to detect packets that are duplicated within 100 ms of each other, so as to avoid double processing which was causing bugs, but also so as not to reject new and relevant information. Consistency seems consistent with and without it being there, so no adverse affects are noted.


Former-commit-id: 046d2385feb406b46db82767a58dc8cea60eead3 [formerly 10edfc06760532e4741377aef42afe5c10515b29]
Former-commit-id: 48c7b7022d10c6a5a5494cedd3e0097aeb3a211e
Former-commit-id: 51ac50d2790e58081950152499854147e5b618b3