Added duplicate packet recognisation
Added a way to detect packets that are duplicated within 100 ms of each other, so as to avoid double processing which was causing bugs, but also so as not to reject new and relevant information. Consistency seems consistent with and without it being there, so no adverse affects are noted.